32->64 ioctls
git-svn-id: http://oss.oracle.com/projects/oracleasm/src/trunk@82 af2e7e45-1c97-403c-8c3e-045742b29be9

+++ kernel/osm.c
@@ -2260,7 +2260,7 @@ static int osmfs_dir_ioctl(struct inode * inode, struct file * file, unsigned in
 					   (struct osm_get_iid *)arg,
 					   sizeof(new_iid)))
 			    return -EFAULT;
-			if (new_iid.gi_version != OSM_ABI_VERSION)
+			if (new_iid.gi_version != (u64)OSM_ABI_VERSION)
 			    return -EINVAL;
 			lock_osb(osb);
 			new_iid.gi_iid = (u64)osb->next_iid;
@@ -2276,11 +2276,11 @@ static int osmfs_dir_ioctl(struct inode * inode, struct file * file, unsigned in
 					   (struct osm_get_iid *)arg,
 					   sizeof(new_iid)))
 			    return -EFAULT;
-			if (new_iid.gi_version != OSM_ABI_VERSION)
+			if (new_iid.gi_version != (u64)OSM_ABI_VERSION)
 			    return -EINVAL;
 			lock_osb(osb);
 			if (new_iid.gi_iid >= (u64)osb->next_iid)
-			    new_iid.gi_iid = 0;
+			    new_iid.gi_iid = (u64)0;
 			unlock_osb(osb);
 			if (!new_iid.gi_iid)
 			{
